LoggerOut can optionally enable p-trx
